[COLAB-2259] Extract methods from ImageDisplayController to ImageDisplayService and LegacyImageDisplayController
